- Split requirements download script into separate download and install scripts

+++ download_requirements.py
@@ -10,34 +10,35 @@ from pathlib import Path
 LIBRARIES_TO_INSTALL = [
     "pydeps",
     "pyelftools",
+    "fabric",
 ]
 
 # The target architecture of your offline device (e.g., aarch64, x86_64)
 # 'manylinux' wheels are widely compatible on Linux.
 TARGET_PLATFORM = "manylinux2014_aarch64"
 
-# Names for the download directory and final archive
 DOWNLOAD_ROOT_DIR = Path("./requirements-downloads")
-INSTALL_ROOT_DIR = Path("./requirements")
 # --- End Configuration ---
+
+def log(string: str = ""):
+    print(string, flush=True)
 
 
 def run_command(command: str):
-
-    process = subprocess.run(
-        command,
-        check=True,
-        capture_output=True,
-        text=True,
-        encoding='utf-8'
-    )
+    log(f"Running command:")
+    log()
+    log(f"{command}")
+    log()
+    process = subprocess.run(command, check=True)
+    log()
 
     return process
-    
+
 
 def main():
+    log(f"Parsing arguments...")
     parser = argparse.ArgumentParser(
-        description="Download and install Python libraries into a portable directory for offline use."
+        description="Download Python libraries into a portable directory for offline use."
     )
     parser.add_argument(
         "python_version",
@@ -48,18 +49,20 @@ def main():
 
     py_version = args.python_version
     # Convert "3.12" to "cp312" for the ABI tag
-    abi_tag = f"cp{py_version.replace('.', '')}"
+    major_minor = "".join(py_version.split('.')[:2])
+    abi_tag = f"cp{major_minor}"
 
     # Get version-specific directories
     download_dir = DOWNLOAD_ROOT_DIR / py_version
-    install_dir = INSTALL_ROOT_DIR / py_version
 
     # Create the required directories
+    log(f"Setting up download root directory...")
     DOWNLOAD_ROOT_DIR.mkdir(exist_ok=True)
-    INSTALL_ROOT_DIR.mkdir(exist_ok=True)
-    
+
     # Delete old downloads for this version
-    shutil.rmtree(download_dir)
+    log(f"Setting up download directory for version \"{py_version}\"...")
+    if download_dir.exists():
+        shutil.rmtree(download_dir)
     download_dir.mkdir()
 
     # Build and run the pip download command
@@ -73,22 +76,8 @@ def main():
         "-d", str(download_dir),
     ]
     run_command(download_command)
-    
-    # Delete old installs for this version
-    shutil.rmtree(install_dir)
-    install_dir.mkdir()
 
-    # Install the packages from the cache into the final target directory
-    install_command = [
-        sys.executable, "-m", "pip", "install",
-        *LIBRARIES_TO_INSTALL,
-        "--no-index",
-        f"--find-links={download_dir}",
-        f"--target={install_dir}",
-    ]
-    run_command(install_command)
-
-    print("Done!")
+    log("Done!")
 
 
 if __name__ == "__main__":
